LayoutTests/printing/page-break-always-for-overflow-expected.txt

Test page-break-{before,after}:always for overflow:{visible,hidden,scroll,auto} elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS: page number of "test-page-break-before:always-for-overflow:visible" is 1
PASS: page number of "test-page-break-before:always-for-overflow:hidden" is 2
PASS: page number of "test-page-break-before:always-for-overflow:scroll" is 3
PASS: page number of "test-page-break-before:always-for-overflow:auto" is 4
PASS: page number of "test-page-break-after:always-for-overflow:visible" is 4
PASS: page number of "test-page-break-after:always-for-overflow:hidden" is 5
PASS: page number of "test-page-break-after:always-for-overflow:scroll" is 6
PASS: page number of "test-page-break-after:always-for-overflow:auto" is 7
PASS: page number of "test-last-page" is 8
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
